Add initial structure
This adds initial structure and also adds initial version
of Value and Environment models.

Environment is a CI model that represents an actual CI environment
or domain that can include multiple CI systems.

Value is a model used for storing data related to the attributes of
the different CI models.

Change-Id: If7ca3d6d679e3804508f109605a24363d9ff61f4