Add configuration handler
This change adds an initial class for handling configuration
definition and loading.

In addition, basic parser was added with support for changing
the configuration file path and enabling debug to see more
detailed output.

An empty sub-parser was added. In future changes, this sub-parser
will be populated with information loaded from the configuration
and the different CI entities.

Finally, README was updated with instructions on how to install
and run CIbyl