Mark attributes that the app should populate
The sources need to know which attributes of the entities to
populate. In order to achieve that, 'populate' attribute was
added to ValueInterface class.

Once the user passed all the arguments, an iteration over all
the entities starts and marks which attributes to populate based
on the arguments the user has used.