Associate model arguments with functions
Once an argument is associated with a function,
we can execute that function if the user has used
the argument. For example: '--jobs' can trigger the
method Source.get_jobs().

In order to not run all the functions associated with
the arguments the user has used, we've also introduced
a level variable to argument class.
The level variable will be used to execute the function
which is in the most advanced level.

Environment
   System
      Jobs
        Job -> Run this function and not any of the previous
               levels, even if user has used them.

Finally, also introduced a 'populated' variable
mark whether information in regards to certain used argument
was pulled from the source to give the user an indication
as to what the source has pulled.