Associate model arguments with functions
This change expands the attributes of "Argument" class in order to:

1. Associate arguments with functions. This will allow us
to invoke a function when an argument is used. So --jobs
argument will invoke Source.get_jobs()

2. In order to not run every function associated with
the different arguments the user has used, this change also
introduces a "level" attribute to determine the level
of the model.

Environment
   System
      Jobs
        Job -> Run this function and not any of the previous
               levels, even if user has used them.

The reason level is part of Argument and not the Model itself
is because models relationship differs based on the system
(Zuul -> Pipline -> Job vs. Jenkins -> Job for example. Job level
is different for job in every system).

3. Finally, also introduced a 'populated' variable to
mark whether information in regards to certain used argument
was pulled from the source to give the user an indication
as to what the source has pulled. This can be also used in the future
to mark failure/success for a source and move to the next one in the
list.