Add configuration options for logging
Adds two cli arguments to configure logging. log_mode configures where
to write the logs (terminal, file or both) and log_file configures the
name of the file to write them. This arguments are read from sys.argv
before the parser is extended in order to configure the loggers and have
them use the same configuration.