Reworking environment creation.
This rework introduces an adapter that allows translating input data to the needs of each source constructor.

Each source may have its own needs and most often than not they are not going to share the same constructor signature. For example, the Zuul source has a 'cert' parameter that the others do not.