Consider system type when creating ci environment
This change connects the ZuulSystem model to the rest of CI models.
Since Zuul systems will have a different model hierarchy, we need to
introduce the Pipeline model in the general system API. The attribute
will then be ignored for all systems that are not zuul, but will
provide the correct interface for zuul.

The API for System defines first the --pipelines arguments and then
--jobs. This has two reasons:
1. In this order, when extending the parser, jobs will have a deeper
   level than pipelines, providing the right entrypoint for both zuul
   and non zuul jobs.

2. The general System will have a jobs attribute. Since the API is used
   by Model to set the attributes, we need to duplicate --jobs to ensure
   that non zuul systems have a jobs attribute.