Duplicate APIs for systems
With this change the System API will be used as reference to extend the
parser, which will give all models the right level. Additionally a new
class is introduced: GenericSystem which has it own API, without
pipelines and with jobs. Therefore, ZuulSystem and GenericSystem
systems will be usable and provide a model for systems that have
pipelines and jobs as a top level model, respectively. The System class
will act as an abstract class that will provide common functionality and
the API to extend the parser.