Allow user to view method without config file
Before this change, if user tries to run application
executable with -h or --help and the configuration file
doesn't exists, it will fail.

Normally, this shouldn't happen as the user can provide a
different path for the configuration file with a dedicated
argument. But, in order to know that in the first place,
user should be able to run -h or --help without failing.
This is what this change does. It checks if user has specified
-h or --help and instead of trying to load the configuration
and create environment instances, it prints the help message.

In addition merged the methods that iterate over sys.argv as
they are essentially the same and this allows us to avoid iterating
sys.argv twice.