Add plugin mechanism
By default, Cibyl makse use only of CI models
but in most cases, builds are associated with project/product
information which isn't reflected in CI models by default.

This change adds the plugin mechanism in order for users
to be able to associate CI models with Product models.

The mechanism loads dynamically the plugin from plugins
folder based on the plugin name, in order to avoid
maintaining registry in Cibyl of plugins and allow anyone
to write its own custom plugin.

The interface of a plugin is a simple one - it should
implement one method called "extend" which extends the
CI models with product-based models.

This change also fixes a bug where the parser would
only extend its arguments if the inspected model API
specified arguments (rather than checking if the API of
the attribute type has arguments).