Rename systems and leave jobs-based API as default
Move all system creation into a SystemFactory class. We now can create
two kind of systems: JobsSystems, which model CI with Jobs as the top
level model and PipelineSystems, which model CI with Pipeline as the top
level model. By default the API will have jobs as the top level module,
since most CI systems will use that. But if a PipelineSystem is
needed, the API will be changed to reflect that and create the
correct hierarchy.