Introducing user and token to urls generated by the Jenkins source.
I have modified how queries are generated by the Jenkins source so that it includes the provided username and token. This will allow us to communicate with hosts running on the 'http' scheme.

I am not very proud of the way the URL is generated, but have not found a library to make it easier. I am open to suggestions because I feel that URLs is something we are usually have to work with.