Creating e2e testing environment and a sample of how a test would look like.
E2E testing consists on asserting the complete flow of one of the app's functionalities. In this case, I have run a Jenkins instance on a Docker container and had Cibyl perform a query over it in order to check that the Jenkins source is working right and the answer is as expected.

E2E tests do not form part of the tox pipeline yet and must be explicitly called with: `tox -e e2e`. This is due to two reasons: first, these tests usually take longer to execute and will slow down the pipeline; second, right now they require some set up that is not automated yet.

In order to run the test cases, you will need to install the Jenkins image first:
`docker build -t cre/jenkins images/jenkins`