Print available systems if non-existing is passed
If the user asks for a non-existing system or environment, the exception
will print all the available systems or environments. This removes the
NoValidEnvironment exception, that will no longer be needed, but keeps
the NoValidSystem one, that might be necessary if user queries by system
type and no system is found.