Introduce speed index mechanism
Right now, users will get an error if they use more
than one source that supports the argument they have used
as part of the CLI usage.

This far from optimal user experience because we don't
want to:
1. Stop cibyl execution just because user has choice, while
   we know exactly what the user is interested in
2. Let the user figure out alone which source is the best
   for each type of argument or query perfomed

This change tries to change this behavior and introduce
a mechanism which will allow developers to quite easily
specify the relative speed of not only the source they
are developing, but the speed of each argument in each
function used during the execution.

The mechanism will aggregate the score from each argument
(assuming it exists in the first place) and if the overall
score is bigger than the score of another source, then that
source will be the preferred one for the execution.

If no speed data is available, then the choice will be arbitrary,
but will not stop the execution anymore.