Add support for --topology argument in jenkins
Modify the get_deployment method of the jenkins source to get topology
information. The information is obtained from the job name or from
artifacts. If many jobs are requested, it falls back to parsing the job
name. If few, it request some artifacts from the last build and tries to
find the information there. This is not completely robust, since
different jobs have different artifacts. If a job has no last build or
the artifacts used are not found, it will also fallback to extracting
the information from the jobs name.