Introduce base speed for speed index
This change introduces a base speed index for each source method, which
can be complemented for some attributes. For example, in the case of
builds for the jenkins source, all attributes have roughly the same cost
except for --last-build, which makes the query faster, thus it adds one
speed point to the index. If an argument makes the query slower, it can
add a negative number to the speed index to take into account. Also, the
default value for the speed index is set to 0.