More robust job filtering in jenkins source
The is_job function in jenkins source expected the jenkins jobs to have
the string "job" in their "_class" argument. However jenkins jobs can
also be of type "hudson.model.FreeStyleProject". With this type, the
function checks that the jenkins object is not a folder or a view, so
freestyle projects are not filtered.

Solves OSPCRE-360