Clearer message for no sources found with --sources
Until now, the environments and systems would be filtered according to
the values of --env-name, --systems and --sources arguments, presenting
a unique error message if no valid system was found. This was confusing
if the source of error was the --source value. With this change, the
checks for --systems and --sources are separated and provide to
different  messages in case of error for better user experience.