Allow disabling systems in configuration file (#179)
* Refactor function to check environments in validator

There was some duplicated code looping over the environments that could
be extracted into a function

* Allow disabling systems in configuration file

Add a 'enabled' parameter for systems in the configuration file. If
systems are not enabled, they will be ingnored completely. This
parameter can be overriden if the system is specified via the --systems
cli argument. If no enabled system is found, a specific error is raised.