Skip source setup when the source is disabled
This makes sure that we don't setup sources which
are disabled in the configuration.

This also introduces a couple of additional minor changes:

* Don't log every duplicated argument. It might look/work
reasonably well with one environment, maybe two, but
when including multiple environments, the amount of lines
we get about duplicated argument is just too big.
* Don't log "setting up source" if the source isn't really
setting up anything. Keep the method empty as an interface
reference for developers.
* Color systems, sources, ... exceptions to highlight
what should be used