Introduce --spec argument for Jenkins source
This change instroduces the --spec argument for the Jenkins source,
which will print the full Openstack specification for a job. It
also adds support for filtering using the --container, --packages
and --services arguments. With this change, informations about
containers, packages and services will only be queries if the respective
arguments are used, givent that it takes significantly longer that other
deployment details.