Add intemediate layer bewtween Source and sources
This changes introduces a new class ServerSource which inherits from
Source and from which the Jenkins, ElasticSearch and Zuul sources will
inherit. At the moment, this class will only contain a method to check
that the required arguments are provided to get_tests, but in the future
it might include additional shared functionality between the three
sources.