Support jobs_scope argument for jenkins and elasticsearch
This change adds the jobs_scope attribute to use with jenkins and
elasticsearch sources. The attribute is part of the system
configuration, so it introduces a new method of System to export system
attributes that might be relevant for sources. This change also modifies
the run_query main loop to create a new SourceStoreMethod per system,
since the assumption that sources will have unique names is not
correct.