Raise exception if configuration file is empty (#214)
Yaml returns None when parsing an empty file. This is checked after
loading the configuration and an exception is raised. This exception
follows the same behaviour as the case of no configuration file, if the
user passed the -h option, the help is displayed, if not, the exception
is raised. Adds also a few test to avoid regressions.