Support passing job name to --spec argument
Support calling --spec job_name without any reference to the --jobs
argument. Values passed to the --spec argument should be an exact match
and not a regex pattern like --jobs. If more than one job is
requested an  error is raised, and if the query does not turn up any
result a different error is raised.