Set granular exception for source misconfiguration
Before this change, if user made a mistake in the sources section
in the configuration file, it would throw a generic exception
about invalid configuration. With this change applied, the user
will now see the invalid source configuration and also the link
to the documentation will be provided for more details.

In addition, consolidated the ConfigurationNotFound exceptions
into one generic exception, and created a separate exception
for empty file.