Fix issues in run_query due to many args and envs
Revert a recent change that swapped the order of the systems and
arguments loops in run_query method of the Orchestrator. The argument
loop must be the outermost one. Otherwise different systems might
perform different queries if, for example, one of the systems does not
have any source that supports the argument with the highest level. This
change also updates the last_level variable, to ensure that queries like
cibyl --spec --jobs do not query for jobs, since --spec is the argument
with the highest level, and --jobs acts simply as a filter.