Add exception for non-supported source key
Before this change, if the user didn't use a valid key
in one of the sources' configuration sections, the exception
thrown was generic and didn't tell the user much on what
exactly the issue is.

With this change, the user will know which key is not supported
and for what type of source.