Add exception for non-supported key in system config
Before this change, if the user didn't use a valid key
in one of the systems' configuration sections, the exception
thrown was generic and didn't tell the user much on what
exactly the issue is.

With this change, the user will know which key is not supported
and for what type of system.