Add ironic options to spec (#242)
* Add ironic_inspector use to spec in jenkins source

Support querying by use of ironic inspector in get_deployment and add
it to the spec option.

* Categorize the printing of deployment properties

Show categories like Network, Storage, Ironic, etc. to separated the
depldeployment properties when printing the results

* Add cleaning network option to spec

This commit adds the cleaning network option to the spec
and provides support for it in the Jenkins source.