Adjust spec output based on verbosity level (#245)
Modify spec output based on verbosity level

The challenges with the output of --spec argument are:

* Keep it consistent so it can be used for purposes
  like jobs comparison
* Keep it concise so user gets valuable information

Before this change, the first point of keeping it consistent
and full was fulfilled, but with providing consistent output, it
slowly becomes impossible to also make the spec output valuable,
as users can see in most of the lines "N/A" if the information is
not available. Perhaps this point isn't critical right now,
but it becomes more critical as more properties are being
added to the spec.

This changes tries to deal with it by balancing between consistent
output and valueable output: if user runs simply --spec, only
available information will be displayed. If user runs --spec -v,
then a full spec will be printed.

This way users can have short concise output, but the same time
ask for more than that and also have full output for purposes like
jobs comparison or seeing what full spec is all about.