Ensure that system constructor arguments are set
Removing the kwargs in the abstract System class prevented some __init__
arguments of JobsSystem and ZuulSystem to be considered in the
constructor. Those arguments that were in the API but were not passed to
the parent's __init__ were overriden there. For example, passing jobs to
a new JobsSystem or tenants to a new ZuulSystem would not work as
expected, since when the Model constructor would be called it would
override the passed values and set them empty. To ensure that regression
occurs, this change also introduces unit tests that cover the
constructor behaviour and the NonSupportedSystemKey exception.