Add openstack cli arguments to get_query_type
In order to introduce the plugin arguments in the get_query_type
function, it will use a QuerySelector class, where functions to check
the query type will be injected by the plugins. The largest value
provided by either the core cli argument or the ones from the plugin
will be returned. This changes maintains the same interface for the
get_query_type and uses the class only internally, to avoid breaking any
existing code.