Switch to query and publish for each system
The current behavior is to first query all the systems
specified in the configuration file and only then publish the
results. The problem with this approach is that the user sees
nothing until all the queries are performed. Perhaps for one
or two systems that's not too bad, but for a dozen of systems
that becomes a bad UX.

This change modifies this behavior to publish the results for
each system, allowing the user to see some output, even if
it didn't finish to query all systems.

In addition, before each query, a message is logged to the
screen, letting the user know that a query started.

This change also modified the publisher so publishing
can done at different levels and no longer it's assumed
that publishing starts always with an Environment.