Allow to disable environments (#267)
Add support for disabling environments

Right now the key "enabled" is supported only
in systems and sources, allowing to disable them
and ignore them while running the application commands.

This change adds the same for environments. Allowing users
to ignore an entire environment with "enabled: false"
configuration.

In addition, fixed a bug with validating environments
specified by the user.