Create a status bar (#302)
This patch introduces a status bar. The status bar can be used to
ensure a user that the program is still working.

The status bar is implemented as a separate thread that starts before
a given code section. During the execution of the given section,
an animation is displayed. Once the section is finished, the thread ends
and so does the animation.