Reuse spec jobs check in elasticsearch
This change introduces a new change to the ServerSource via the
openstack plugin to ensure that all three sources that will implement
the option show the same behaviour. The method checks that only one job
is provided, either with the --spec option or the --jobs one. If there
is more than one option an exception is raised. If there are no matching
jobs, a different exception is raised. The method is not added to the
Zuul source since the get_deployment method is still under development.