Update documentation with entities drawing
Added a drawing and a short explanation to Cibyl's index
page that describes what are environments, systems and sources.

The purpose of the drawing and the explanation is to set the right
mindset of a user on how Cibyl should be used when the user is reading
the documentation for the first time.