Only print topology and nodes when asked in Jenkins (#334)
Some arguments like --spec, or combinations like using --packages or
--containers with many jobs would cause cibyl to output unasked
information like the topology or the list of nodes. This change ensures
that the information on topology and nodes is only added when the user
asks for it with the --topology or --nodes arguments, or the --spec in
the case of the topology.