Use logger from core jenkins source in plugin extension (#338)
Due to the way the plugin source extensions are loaded, the loggers
defined there would not inherit the logger configuration defined for the
application. This change importes the logger from the core source in the
extension to ensure they have the same formatting. It also removes the
defined logger from the elasticsearch extension since it would have the
same problem but was not used. If logging is required there, it should
follow the same pattern and import the one from the core source.