Construct the queries graph and use it to select args
Construct a graph with relationships between sources query methods. The
graph is constructed as arguments are added to the parser and requires
the orchestrator to  deduce the relationship between two query methods.
It does so by considering the func attribute from the arguments when an
API entry points to a model that also has an API.

For example, when the extend_parser method encounters the --builds
argument in the Job API, it remembers that first it had encountered
a --jobs argument with get_jobs as the query_method. When the --builds
argument is added to the parser, it is also added to the queries graph
connecting get_jobs and get_builds.

Since get_jobs is the parent of get_builds, it assumes that in order to
resolve get_builds, get_jobs will also be called internally, so there is
no need to call it explicitely.

Arguments that have a func but no parent query method are assumed to
be roots in the graph (e.g. get_jobs for jobs systems and get_tenants
for zuul systems).

The input arguments are then filtered using this graph when running
the query. Instead of simply using the level as an heuristic, the new
method selects the argument at the highest level for each branch of
the graph, which solves the issues of some queries being ignored.