Rework proposal to avoid adding parent_func attribute
Modify the proposal for query selection to avoid explicitely adding the
parent_func attribute to the arguments. The modification deduces the
parent_func attribute from the query method found when visiting a new
API. For example, when the extend_parser encounters the --builds
argument in the Job API, it remembres that first it had encountered
a --jobs argument with get_jobs as the query_method. When the --builds
argument is added to the parser, it is also added to the queries graph,
connecing get_jobs and get_builds.