Add type hints and return types to methods in utils package (#366)
* Add type hints and return types to methods in utils package

- Add type hints for utils subpackage
- Remove type hints in docstring

Note: It seems that we aren't able to specify the return type for a self class if Python version <= 3.11 (from typing import Self - https://peps.python.org/pep-0673/). I've seen the way to specify the same class is in a string way. If the self class is e.g. Production then the type should be 'Production'.

* Add type hint to Color class methods

* Fix type hints to the right ones

* Fix flake8

* Fix type hints

* Fix type hint

* Remove unneeded docstring types and add missing type hint

Co-authored-by: jgilaber <jgilaber@redhat.com>