Refactor extend_parser to better group arguments (#389)
The extend_parser method of the orchestrator would not always group
arguments to the model they point to. For example, the --builds argument
was listed under the Job model, and the --tests argument under the
Build section. This changes groups all arguments with the model they are
related to, and makes sure that all API encountered are explored. That
is, if model A has in its API an entry for model B, but no arguments
associated, it still would recursively explore the API of model B.