Adjust speed_index for Jenkins source
In the initial release of Cibyl the Jenkins source supports querying
for more information. Because of that, we want Jenkins to be the
preferred source over Elasticsearch in certain cases.

This change modifies the speed_index for Jenkins source so that 
the Jenkins source is preferred over the Elasticsearch source in certain 
cases.