Fix bug unable to find tripleo schemas outside repo
When running a query loading the tripleo json schemas from a path
outside the source code folder. The default path to look for those json
schemas was specified as a relative path, which if run from a random
location would not find the schema files. This was solved by changing
the paths and making them relative to the tripleo package path, so it
can be located from any working directory.