Move source setup inside try/except in orchestrator
The sources setup method was called outside a try/except block,
which would cause to skip the rest of the sources if an error ocurred
during the execution of the setup. This changes moves the call to the
setup inside the same try/except block where the call to the query
method occurss, so any SourceException raised during the setup will be
caught and the next source (if available) will be called.