Add spec subcommand and update documentation (#456)
Add a new spec subcommand that should replace a call like cibyl query
--spec job-name. The same command would be now cibyl spec job-name.
Functionality stays the same, that is, calling cibyl spec without
a job-name will fail, and calling it with multiple job-names will also
fail. This change also removes the --spec argument, updates the
mentions to it in the documentation to new subcommand and reactivates
a couple of tests that were skipped in the past.