Add mechanism to add subcommands from plugins
Implement a mechanism to provide product-specific subcommands from the
plugin themselves. At the moment, there is only one example of this, the
spec subcommand provided by the openstack plugin. The plugins that add
subcommands must provide a list of functions that take an ArgumentParser
object as argument and add the subpbarser there. This list is collected
when the plugins are loaded and passed to the add_subparser method of
the Parser class to form the complete cli.