Add e2e jenkins testing environment for openstack (#472)
Add a e2e testing environment for the Openstack extension of the Jenkins
source. This requires not only setting up a Jenkins mock instance, but
a mock of the logserver where the deployment information can be read,
and linking to this log server from the description of  the Jenkins
build.