Add separate publisher for JSON output
Producing JSON output for multiple environments requires accumulating
the environments string representation and print it all at once, after
all queries have finished. On the other hand, when printing to screen,
we want to show the output as soon as each environment is queried,
so this change introduces two Publisher classes, one for each use
case.