Make parsing of the plugin argument aware of subcommands (#486)
The parsing of app-level arguments done before the full parsing did not
take into account the subcommands. This creates a problem if plugins are
specified with the -p flag, since it would read plugins until another
argument was found, but could not distinguish between plugin names and
subcommands. This commit fixes the problem by keeping a list of all
available subcommands and using it to check which arguments are plugin
names.