Add docs page for cibyl output for user and devs
Add two new pages to the documentation explaining the cibyl output. One
page is from the point of view of the user, explaining output formats
and how to select them. The other is from a develepor perspective,
trying to provide a summary of the architechture used for the components
responsible for output creation.