Fix two bugs related to spec in Jenkins sources (#518)
Download the last successful build from jenkins, instead of the
last completed build which will be sure to provide the logs with the
most information. Also, this change also fixes a bug in the regex that
read the logs url from the build description. The previous regex was too
general, and in cases where more than one link was present in the
description, then all of them were included. The regex is restricted
further so than only one url will be captured.