Adapt get_builds to new elasticsearch index (#561)
Adapt the get_builds method of the Elasticsearch source to work with the
new index. Since we now have unique documents for the (job, build) pair,
get_builds now does not call get_jobs explicitely, but rather makes
a new query getting also the job information and creating the
corresponding models, avoiding an uneeded query that adds a large amount
of time.