Adapt get_deployment to new elasticsearch index (#566)
Adapt the get_deployment method of the Elasticsearch source to work with the
new index. Since we now have unique documents for the (job, build) pair,
get_deployment now does not call get_jobs or get_builds explicitely, but
rather makes
a new query getting also the job information and creating the
corresponding models, avoiding an unneeded query that adds a large
amount of time.